+++ common/src/main/java/net/merchantpug/bovinesandbuttercups/mixin/client/GuiGraphicsMixin.java
@@ -7,6 +7,7 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.network.chat.Style;
 import net.minecraft.util.StringUtil;
 import org.spongepowered.asm.mixin.Mixin;
@@ -23,7 +24,15 @@ public class GuiGraphicsMixin {
         if (!list.isEmpty() && list.get(0).contains(BovineEffects.LOCKDOWN.get().getDisplayName())) {
             List<Component> newList = new ArrayList<>(list);
             newList.remove(1);
-            Services.COMPONENT.getLockdownMobEffects(Minecraft.getInstance().player).forEach(((statusEffect, duration) -> newList.add(statusEffect.getDisplayName().plainCopy().append(" ").append(StringUtil.formatTickDuration(duration, Minecraft.getInstance().level.tickRateManager().tickrate())).setStyle(Style.EMPTY.withColor(ChatFormatting.GRAY)))));
+            Services.COMPONENT.getLockdownMobEffects(Minecraft.getInstance().player).forEach(((statusEffect, duration) -> {
+                MutableComponent component = statusEffect.getDisplayName().plainCopy().append(" ");
+                if (duration == -1)
+                    component.append(Component.translatable("effect.duration.infinite"));
+                else
+                    component.append(StringUtil.formatTickDuration(duration, Minecraft.getInstance().level.tickRateManager().tickrate()));
+                component.setStyle(Style.EMPTY.withColor(ChatFormatting.GRAY));
+                newList.add(component);
+            }));
             return newList;
         }
         return list;