+++ neoforge/src/main/java/net/merchantpug/bovinesandbuttercups/BovinesAndButtercupsNeoForge.java
@@ -343,6 +343,8 @@ public class BovinesAndButtercupsNeoForge {
                 Services.COMPONENT.getLockdownMobEffects(event.getEntity()).forEach(((statusEffect, integer) -> {
                     if (integer > 0) {
                         lockdownEffectsToUpdate.put(statusEffect, --integer);
+                    } else if (integer == -1) {
+                        lockdownEffectsToUpdate.put(statusEffect, -1);
                     }
                 }));
                 Services.COMPONENT.setLockdownMobEffects(event.getEntity(), lockdownEffectsToUpdate);